refactor(tests): include lib using bindings
